requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
deta==1.1.0
discord==2.2.2
discord.py==2.2.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
sniffio==1.3.0
yarl==1.9.2
